function/requirements.txt
azure-functions

azure-identity>=1.12.0
azure-core>=1.26.0


python-dotenv>=0.19.0
azure-functions
azure-search-documents==11.4.0b7   # preview that includes agents/knowledge features
azure-identity>=1.16.0
openai>=1.25.1
python-dotenv>=1.0.1
requests>=2.31.0